TST: move misplaced to_datetime test (#32273)
